refactor(chat): extract inline logic into custom hooks and subcomponents
- Decompose ChatView into smaller custom hooks:
  useChatInference, useChatMessageActions, useChatModelPath,
  useChatModelStatus, useChatPopupPanel, useChatTextareaResize
- Split monolithic MessageList into dedicated message components:
  MessageBubble, StreamingBubble, MessageActions, MessageAvatar,
  ToolCallSection, EmptyPlaceholder
- Extract input area into ChatInputArea with sub-components:
  ActiveToolsDisplay, ChatDocsRow, ChatToolbar, ModelLoadingIndicator
- Add ChatPopupPanel for knowledge-base and history panels
- Add shared UI primitives: PanelIconBtn, ToolbarToggle
- Add chat types module
- Simplify useConversationMessages to work with new architecture
