+++ client/src/components/chat/types.ts
@@ -1,0 +1,6 @@
+export type ChatPanel = "knowledge" | "history" | null;
+
+export interface ActiveDoc {
+  id: number;
+  name: string;
+}
